soil moisture sensor & IR sensor
since we working on this project to explore Mars. So here Soil Moisture sensors and IR sensors play a major  role in the exploration
as soil moisture sensors detect moisture in the Martian soil which will help us to know whether a living being might survive or not on Mars.
And IR sensors play a crucial role in this. It helps in studying atmospheric analysis, Terrain mapping, Navigation and Obstacle Avoidance